stop threading, draw at framerate
Threads should be handled by the modules themselves.
I see a case where the number of colors makes a module impossible to
run at the target FPS and might warrant a separate thread at the top
level, but thats only practical when the draw time is long enough for
the module thread to be able to generate another frame. For now, its not
worth implementation but when there are more costly draw methods and
potentially high numbers of colors, it could net some nice performance.